requirements.txt
fastapi
uvicorn
httpx
PyNaCl
qrcode[pil]
python-multipart
python-dotenv
pydantic-settings
jinja2
structlog
